Episode 25

+++ app/Http/Controllers/ThreadsController.php
@@ -120,6 +120,8 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize('delete', $thread);
+
         $thread->delete();
 
         if (request()->wantsJson()) {
